Initial project setup and accelerated regex filter
Initial setup is pretty normal for an Elasticsearch plugin.

The docs best describe the accelerated regex filter:
The ```source_regex``` filter is designed to quickly perform arbitrary regular
expression searches against the source documents.  It uses an ```nGram``` index
to select only documents that might match and then runs the regular expression
against those documents.  In simple local testing that is an order of magnitude
faster (50ms -> 5ms) but it ought to be much much better on real, large
documents (minutes -> tens or hundreds of milliseconds).

Its capable of accelerating even somewhat obtuse regular expressions like
/a(b+|c+)d/ and /(abc*)+de/ using the algorithm described [here](http://www.pgcon.org/2012/schedule/attachments/248_Alexander%20Korotkov%20-%20Index%20support%20for%20regular%20expression%20search.pdf).
This filter can also limit the execution cost of even regular expressions that
can't be accelerated by that method by limiting the number of documents against
which a match attempt is made.

Change-Id: I9b82ea532cf846e6129570a21576aebf3163f3eb