Add field_value_factor_with_default
This is a backport of an elasticsearch feature to add the "mising" field to
field_value_factor. That field contains a value that is used by the
field_value_factor when the document is missing the field.

Alomst a backport strait from the master branch but required some small
tweaks.

Change-Id: I8b11eee35cf0579e3baba5ea051084306333a455