Document level noop detection with a version field
Elasticsearch has it's own versioning support, which is quite useful,
but unfortunately it is not available when using the _update API which
the super_detect_noop script is executed through. Add a handler for each
document to contain a version field, and handle it similar to
external_gte (but using a document value, rather than the elasticsearch
internal version).

Bug: T144039
Change-Id: Ib697b5e5fd11c98eb084631fecaf4a1b9ad10dad