Speed up the regex's recheck phase
Here are timings for 1000 iterations of rechecking the Barack Obama and
Rashidun Caliphate articles from English wikipedia:

              slow case insensitive took 12331 millis to match /\[\[Category:/
  non backtracking case insensitive took 4525 millis to match /\[\[Category:/
   case converting case insensitive took 4112 millis to match /\[\[Category:/
                slow case sensitive took 11099 millis to match /days.+and/
    non backtracking case sensitive took 3431 millis to match /days.+and/
     case converting case sensitive took 579 millis to match /days.+and/
              slow case insensitive took 9288 millis to match /cat/
  non backtracking case insensitive took 2893 millis to match /cat/
   case converting case insensitive took 103 millis to match /cat/
                slow case sensitive took 9170 millis to match /\[\[Category:/
    non backtracking case sensitive took 1925 millis to match /\[\[Category:/
                slow case sensitive took 6178 millis to match /cat/
    non backtracking case sensitive took 47 millis to match /cat/
                slow case sensitive took 7669 millis to match /days.+and/
    non backtracking case sensitive took 271 millis to match /days.+and/

These numbers mean:
1. For case sensitive queries the recheck phase gets somewhere between an
order of magnitude and 3 times faster. ~2 seconds for 2000 rechecks is still
not great but its better. 47 milliseconds for the same number of rechecks is
zippy.
2. For case insensitive queries the recheck phase gets faster as well but
the depth of the match from the front of the document causes a more pronounced
slowdown. In the worst case the recheck is still twice as fast as it once
was.

This has tests for Greek, Turkish, and Irish which are all special cases
of case insensitivity. I'm reasonably sure Irish has never worked well and
these tests are no exception. The other languages work well.

Change-Id: I0e14d0436b23425bcdbab65e85a5e6d775cf8e5d