Adding project lombok support
Adding automatic generation of getters, setters, equals and hashCode.

Note that your IDE probably needs support to handle lombok correctly. Eclipse
and IntelliJ both have plugin for it. If IDE support is an issue, we probably
do not want to merge this.

See https://projectlombok.org/ for an introduction to project lombok.

Change-Id: I1e5c90f2813f3e2b934df2841e97d1b81c74ee41