Prevent too large boolean queries from being created
With some complex regular expressions the number of ngram automaton transition
is properly limited by max_ngrams_extracted. Unfortunately these transitions
may loop between each others and can cause the boolean expression generator
to create a very large boolean expression (100 extracted ngrams may be combined
into a 30000+ boolean clauses).
This problem was not really noticed before because we used BooleanFilter which does
not re-create/optimize the boolean tree on each segment and remain relatively fast.
By using BooleanQuery we now hit the problem where a complex scorer is created on each
segment (in addition to the BooleanWeight) and the time required to build the scorer
tree is certainly counter productive.
BooleanQueries being limited to 1024 clauses it seems reasonable to also limit
the number of ngram clauses to 1024 by default.

This patch adds a check on the number of clauses in the boolean expression, if this
number exceeds max_ngram_clauses then a flat disjunction with the ngrams will be used
as the approximation. In some case it may still help to accelerate the query (if none
of the ngram extracted is popular). In other cases it will be just like an
unoptimized regex with a small overhead due to useless ngram lookup.

Bug: T133751

Change-Id: If9fc5c95dfcaa33e6bf3e007eb3781b817d4728f