InvalidRegexException extends RuntimeException instead of IllegalArgumentException
Some exceptions are now manually rewritten by elasticsearch
when sent through the Transport layer.
In this case all exceptions that extends IllegalArgumentException will be
written and read as an IllegalArgumentException, in other words the child class
type is forgotten in the transport.
The quick and dirty solution is to switch to RuntimeException for now.
Note that RuntimeException/IllegalArgumentException has a special meaning in reponse
status behavior, the former will trigger a 500 the later a 400.

Relevant commit on elastic size:
https://github.com/elastic/elasticsearch/blob/e7eb9cf4de13620c473512dcbf0259d390ed471d/core/src/main/java/org/elasticsearch/common/io/stream/StreamOutput.java#L472

Change-Id: I4a73cfc66fce76dd6801662d11b014e0e5bb3b00