Workaround timeout limitation
Timeout checks are passive with Lucene, it means that the timeout is
only checked when a doc matches or when we switch to a new segment.
The problem is that the timeout can be very inaccurate when scanning a
large segment where no docs match.
Workaround this issue by a horrible hack:
- Build a fake collector to be able to throw TimeExceededException
  (private constructor)

Sadly it is necessary to explicitly tell the source_regex query
to use a timout. It's impossible to determine if a timeout
has been set in the top level search request.

Added some code to prevent the query from being cached.
We'd like to avoid caching partial results.

Bug: T152895
Change-Id: I1ce0378e475bf5c122cb8586916722992f82bed3