Regex: Add a post analysis after trigram extraction
Depending on the analysis chain some trigrams may be changed by
various filters when generating the trigram index. Unfortunately when
we extract trigrams at query time we do not apply the same filters.
We simply run a String.toLowerCase on the regex expression itself and
then with extract trigram.
In some cases when for instance we use the case folding strategy of
analysis-icu some trigrams may be completely different, e.g. "ß" is
folded as "ss". By applying the same analyzer on trigram at query
time we can are now able to find these terms.

This is still not an ideal solution but it fixes the most obvious
problems. I don't really like the fact that we apply
String.toLowerCase on the regular expression itself, I'd prefer to
stick with a naive lowercase approach applied after trigram
extraction. But this would be a bigger change and won't be comptible
with existing indices.

Bug: T147636
Change-Id: I1c8bf3ec60d16c39bcd29304a9ef5542c3211a02