check import order at build time
Import order changing randomly is painful for code reviews as it adds non
relevant information to diffs.

This configures a checkstyle check that conforms to the default IntelliJ
configuration. It will make the build fail if this import order is not
respected.

It is still the responsibility of each developer to configure its IDE
correctly.

Change-Id: I97bd3ec1fbdcfb7c0b11b82cab7d13f4120ac6cf