Improve regex performance with leading wildcards
Regexp's that have leading wildcards (.*, etc) had pathological
performance with the prior matching algorithm. The added no match test
case takes more than a minute on my laptop to perform a single
iteration.

Since we only care that a match exists, and not where it is, wrap all
regexp's for rechecking in `.*(regexp)`. The resulting automaton can
be applied to the source in a single pass instead of the nested loops
we had before.

The prefixed .* does increase the number of states in the determinized
automaton. Some of our integration test cases had to change as their
examples were too complex. To evaluate if this will negatively effect
production traffic I ran the prefixed and unprefixed versions against
6.5M prod queries.  Of the 6.5M, 2.5k failed with the unprefixed
regexes. Adding the prefix only caused 4 new queries to fail. This
seems a small enough impact to be worthwhile.

Bug: T195491
Change-Id: Iefa4acc4e8b85a456b61babfc9a3de82856d46d8