Add support for elastic 6.3.0
Most notable change is the timeout support on source_regex.
The query format does not change but in order to get a partial
response the timeout set on source_regex must be slightly higher than
the timeout set on the search request.
The new param allow_partial_results must be set to true as well
otherwise elastic will prefer to return a shard failure in case of
timeout.

Completely removed the deprecated max_inspect param of source_regex.

Bug: T198067
Change-Id: I53543f7fafc44a5e514f7b24cab25e5e6e00b2a7