super-detect-noop: Recognize and update subsets within an array
We want to store class predictions from multiple ML models in the
same elasticsearch field. When updating the predictions we need to
replace the stored predictions only for the provided models, leaving
predictions from un-updated models as they were. Accomplish that
with a ChangeHandler that treats the content and updates as a
multilist encoded within a single list.

Bug: T250237
Change-Id: I65432cdacfe5da459178ecb536104b89898dd2cb