Clarify variables in SuperDetectNoopScript::update
This update function is complex and hard to read. Part of that
complexity is the uncertainty of naming. Is `sourceValue` from the
source, or the updateSource?

This is separated from the followup patch that refactors this method
as these old names made even reviewing the following refactor
difficult as it wasn't clear if the correct variables were being
referenced.

* prefix variables as either old or new, depending on their source. This
  mimics how ChangeHandler names these variables as we pass them down.
* Pull newEntry.getKey() into a variable named simply `key`, as it is
  shared between new and old.
* always suffix Object variables with Raw and remove the Raw suffix
  when casting to a more concrete type.

Change-Id: Ia2b4bd3ad8af0b9af09a11802f020324e651183c