Explicitly track nullability in ChangeHandler.handle
Implementations of the interface varied in if they handled nullability
themselves or if they required the NullSafe decorator to function as
a full implementation of the interface. Create a new interface for the
implementations that are not correct when passed nulls, making the
only correct way to use them to wrap them with the null safe handler.

Change-Id: Ic660dbf1c991f4207cd6c911f24a37f678aa9e59