Accept source via script code in addition to parameter.
Motivation: Size estimation of bulk requests does not consider parameters. Since we pass the script source (document) as parameter, the client logic underestimates the actual size of a request. To work around this, we may also accept the document via the script code property, that is taken into consideration for the size estimation.

Change-Id: I0659f6d47964d34553cc01b283e3093e78717a31