Support anchors and short char classes in source_regex
While the initial use case for this query was the full source
text of a wikitext document, that use case has expanded into
applying regex's to the title of a wiki page. In that context
start and end anchors can be a useful addition. The addition of
anchors required extended handling of character classes. While
we were there support was added for \d, \s, and \w shorthand
character classes.

The rewriting implementation is found in the lucene-regex-rewriter
module of wmf-jvm-utils. This shares the implementation between
the source query and the highlighter.

Short char classes are implemented with a simple expansion from
short to full character class syntax. Short char classes can be
used both inside and outside a character class (ex: \d, [^\d]).
The short char classes are the simple expansion, not full unicode.

Anchors are implemented by using utf8 non-characters \uFDD0 and
\uFDD1 as sigils to mark the beginning and end of the string. The
anchors are replaced in both the accelerated part of the query
and the rechecker, making them fully supported in expected use
cases. Acceleration requires a special index time analysis
component, documented in docs/source_regex.md. The query
auto-detects it's usage and provides anchor support if the
underlying ngram field supports it.

Bug: T317599
Depends-On: Ie5e3ad991235fdd0b6efffc9c747b406e52e8e00
Change-Id: Iba064746f175132fd7bd0ac9d4e6a0bffd6a3af5