Pre-apply spotless formatting changes
The next patch changes some files, and spotless is configured
such that it only formats files that are edited in the same
patch. The files in the next patch have never been formatted by
spotless before, thus has a large diff.  Separate that diff out
into this patch rather than confusing the followup with unrelated
changes.

Change-Id: I99106ea43f2ce11977fc7b6b79eb2e1860b82ab3