multilist: Improve support for null source value
When the source value, from the indexed document, was null the
multilist would return the script parameter as the new value to
be indexed, even if it contained a __DELETE_GROUPING__ tag. This
led to indexing those __DELETE_GROUPING__ tags.

Fix by adjusting the null handling routine so that it can
invoke the handler with a non-null value. Now the handler sees
the empty list, deletes nothing from it, and the change is then
recognized as a noop.

Bug: T406205
Change-Id: Ifca9ce2109f714a6e265e759f845b9f26e6497ec