regex: Bump lucene-regex-rewriter to 1.0.7
The lucene-regex-rewriter jar has been updated to resolve a
problem with the way multi-byte unicode expansions were handled,
it has also been updated to add an additional unicode expansion
syntax.

Dependency was updated, and a few tests were added as a basic
verification that the rewritten queries also work as expected
with the trigram acceleration query.

Bug: T403212
Change-Id: Idbbcedb8379c8ee96e30d99db469bb06e062af98